+++ src/__internal__/input-behaviour/index.js
@@ -1,5 +1,0 @@
-export { InputBehaviour, InputContext } from "./input-behaviour.component";
-export {
-  InputGroupBehaviour,
-  InputGroupContext,
-} from "./input-group-behaviour.component";